Create matchLetter
I realized I was trying to compare a character (A) to the variable A that holds a matrix and that just wasn't going to work. 

This function holds a switch menu that checks that user input matches the accepted character values and calls the correct vector matrix if there's a match to the user's input.